% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peaks.R
\name{get_peaks}
\alias{get_peaks}
\title{Title}
\usage{
get_peaks(data, quantile = 0.999, snp.annot = TRUE, rm.loc = TRUE)
}
\arguments{
\item{data}{description}

\item{quantile}{description}

\item{snp.annot}{description}

\item{rm.loc}{description}
}
\value{
description
}
\description{
Description
}
