% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_bim_frq.R
\name{load_bim_frq}
\alias{load_bim_frq}
\title{Title}
\usage{
load_bim_frq(file_path, bim, frq)
}
\arguments{
\item{file_path}{description}

\item{bim}{description}

\item{frq}{description}
}
\value{
description
}
\description{
Description
}
