% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_plot.R
\name{manhattan_plot}
\alias{manhattan_plot}
\title{Title}
\usage{
manhattan_plot(pbs_data, col_name, fig_name = NULL, facet_key = NULL)
}
\arguments{
\item{pbs_data}{description}

\item{col_name}{description}

\item{fig_name}{description}

\item{facet_key}{description}
}
\value{
description
}
\description{
Description
}
