% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot.add.genes.R
\name{mplot.add.genes}
\alias{mplot.add.genes}
\title{Title}
\usage{
mplot.add.genes(data, mplot, maxv_per_chr = T, all_peakv = T)
}
\arguments{
\item{data}{description}

\item{mplot}{description}

\item{gene_col}{description}

\item{merge_col}{description}

\item{by}{description}

\item{label}{geom_text_repel or geom_label_repel}
}
\value{
description
}
\description{
Description
}
