% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot.add.points.R
\name{mplot.add.points}
\alias{mplot.add.points}
\title{Title}
\usage{
mplot.add.points(DT, rolling_DT = NULL, mplot, target_snp, merge_col = "W_ID", uniq = FALSE)
}
\arguments{
\item{DT}{description}

\item{rolling_DT}{description}

\item{mplot}{description}

\item{merge_col}{description}

\item{uniq}{description}
}
\value{
description
}
\description{
Description
}
