% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reynolds_fst.R
\name{reynolds_fst}
\alias{reynolds_fst}
\title{Fst from Reynolds, Weir, and Cockerham (1983)}
\usage{
reynolds_fst(allele.freq, pop1, pop2)
}
\arguments{
\item{allele.freq}{Object from class 'gdinput', obtained by the function "load_bim_frq".}

\item{pop1}{Population ID in the CLST field for population 1.}

\item{pop2}{Population ID in the CLST field for population 2.}
}
\value{
description
}
\description{
Description
}
