% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliding_window_mean.R
\name{sliding_window_mean_dt}
\alias{sliding_window_mean_dt}
\title{Title}
\usage{
sliding_window_mean_dt(dt, col_name, window_size = 20, slide = 5)
}
\arguments{
\item{dt}{description}

\item{col_name}{description}

\item{window_size}{description}

\item{slide}{description}
}
\value{
description
}
\description{
Description
}
