% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot.add.genes.R
\name{mplot.add.genes}
\alias{mplot.add.genes}
\title{Title}
\usage{
mplot.add.genes(data, mplot, gene_col = "GENE", merge_col = "W_ID", by = "CHR", label = "geom_label_repel")
}
\arguments{
\item{data}{description}

\item{mplot}{description}

\item{gene_col}{description}

\item{skip_empty_genes}{description}

\item{merge_col}{description}

\item{by}{description}

\item{label}{geom_text_repel or geom_label_repel}

\item{log_pvalue}{description}

\item{...}{additional parameters for ggrepel functions}
}
\value{
description
}
\description{
Description
}
