% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplot.add.genes.R
\name{mplot.add.genes}
\alias{mplot.add.genes}
\title{Title}
\usage{
mplot.add.genes(data, mplot, maxv_per_chr = T, all_peakv = T)
}
\arguments{
\item{data}{description}

\item{mplot}{description}

\item{maxv_per_chr}{description}

\item{all_peakv}{description}
}
\value{
description
}
\description{
Description
}
