% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbs_fst.R
\name{pbs_fst}
\alias{pbs_fst}
\title{Title}
\usage{
pbs_fst(geno.info, focal, close, outgroup, monomorphic_removal = T, method = "reynolds")
}
\arguments{
\item{geno.info}{description}

\item{focal}{description}

\item{close}{description}

\item{outgroup}{description}

\item{monomorphic_removal}{description}

\item{method}{description}
}
\value{
description
}
\description{
Description
}
