% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reynolds_fst.R
\name{reynolds_fst}
\alias{reynolds_fst}
\title{Title}
\usage{
reynolds_fst(allele.freq, pop1, pop2)
}
\arguments{
\item{allele.freq}{description}

\item{pop1}{description}

\item{pop2}{description}
}
\value{
description
}
\description{
Description
}
